feature/Resolved bugs
