+++ task_application/lib/bottomnavigator.dart
@@ -1,0 +1,84 @@
+import 'package:flutter/material.dart';
+import 'package:task_application/inicio.dart';
+
+class BottomNavigator extends StatefulWidget {
+  const BottomNavigator({super.key});
+
+  @override
+  State<BottomNavigator> createState() => _BottomNavigatorState();
+}
+
+class _BottomNavigatorState extends State<BottomNavigator> {
+  int _selectedTab = 0;
+
+  final List _pages = [
+    const Center(
+      child: Inicio(),
+    ),
+    const Center(
+      child: Text("Projects"),
+    ),
+    const Center(
+      child: Text("Calendar"),
+    ),
+    const Center(
+      child: Text("Menu"),
+    ),
+  ];
+
+  _changeTab(int index) {
+    setState(() {
+      _selectedTab = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        automaticallyImplyLeading: false,
+        backgroundColor: Colors.white,
+        title: const Text(
+          'June 21, 2024 \nHola, David 👋',
+          style: TextStyle(
+            color: Colors.black,
+            fontSize: 20,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        // backgroundColor: Colors.white,
+        actions: [
+          IconButton(
+            onPressed: () {},
+            icon: const Icon(
+              Icons.search,
+              color: Colors.black,
+            ),
+          ),
+          IconButton(
+            onPressed: () {},
+            icon: const Icon(
+              Icons.notifications_none,
+              color: Colors.black,
+            ),
+          ),
+        ],
+      ),
+      body: _pages[_selectedTab],
+      bottomNavigationBar: BottomNavigationBar(
+        currentIndex: _selectedTab,
+        onTap: (index) => _changeTab(index),
+        selectedItemColor: Colors.red,
+        unselectedItemColor: Colors.grey,
+        items: const [
+          BottomNavigationBarItem(icon: Icon(Icons.home), label: "Inicio"),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.file_copy), label: "Proyectos"),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.calendar_month), label: "Calendario"),
+          BottomNavigationBarItem(icon: Icon(Icons.menu), label: "Menu"),
+        ],
+      ),
+    );
+  }
+}
