+++ task_application/pubspec.yaml
@@ -32,7 +32,7 @@ dependencies:
     sdk: flutter
   flutter_slidable: 3.1.0
   dots_indicator: ^3.0.0
-  flutter_speed_dial: ^6.0.0
+  flutter_speed_dial: ^7.0.0
 
 
   # The following adds the Cupertino Icons font to your application.
